refactor(product): extract auto-follow-bottom inset math from stick-to-bottom hook
Splits the manual-only overlay inset tracking, its scrollTop math, and the
scroll-to-bottom callbacks into use-transcript-auto-follow-bottom.ts, a
cohesive sibling of transcript-auto-follow-target.ts. Pure extraction, no
behavior change. Brings use-transcript-stick-to-bottom.ts back under the
400-line soft cap (421 -> 363 lines).
